When detaching, wait for the child process to initialize fully.
This way the pid file can be checked immediately following exit of the parent.

This allows external monitoring applications to watch the daemon without having to add sleep calls to wait for the pid file be written on a busy system.

The wait is implemented using the example from man 2 pipe.

Signed-off-by: Alexander Kanavin <alex@linutronix.de>